[#253] Task B — split step_internal per JEOD integration stage
Carves the ~1,000-line `step_internal` orchestrator into per-stage
sibling files under `simulation/step/`. Each file reopens
`impl Simulation` and exposes one `pub(super) fn` per stage; the
top-level `step_internal` in `mod.rs` becomes a 25-line orchestrator
that calls each in order.

## Layout

```
simulation/step/
├── mod.rs          ~150 L  // step, step_internal (orchestrator),
│                            // step_n, step_until, frame_origin
├── ephemeris.rs    ~165 L  // stages 2 + 2b: planet-fixed rotations,
│                            // frame-tree sync, DE4xx source positions
├── environment.rs  ~115 L  // stages 4 + 4b + 5: gravity (Newtonian +
│                            // PPN relativistic), atmosphere
├── interactions.rs ~180 L  // stage 6: drag, SRP, gravity-gradient
│                            // torque (returns sun_pos / moon_pos for
│                            // stage 9 to reuse)
├── integrate.rs    ~620 L  // stages 7 + 8 + 8b: force collection,
│                            // RK4 integration with contact-pair
│                            // coupling, frame-switch handling.
│                            // Stays one file because the contact-pair
│                            // RK4 branch interleaves stage-7 force
│                            // collection with the per-RK4-stage
│                            // derivative recomputation.
└── derived.rs      ~105 L  // stage 9: orbital elements, Euler
                             // angles, LVLH, geodetic, solar beta,
                             // earth lighting
```

## State flow between stages

`step_internal` threads two pieces of intermediate state across
stage methods:

- `body_integ_origins: Vec<(DVec3, DVec3)>` — per-body integration-
  frame origins resolved against the root frame. Computed once in
  `step_internal` (it's a frame-tree query that doesn't change
  between stages 4 and 8). Passed to `update_environment(...)` and
  `run_integration(...)` as `&[(DVec3, DVec3)]`.
- `(sun_pos, moon_pos): (Option<DVec3>, Option<DVec3>)` — Sun and
  Moon inertial positions, resolved once at the top of stage 6 and
  reused by stage 9's solar-beta and earth-lighting blocks.
  Returned by `compute_interactions()`, passed to
  `compute_derived_states()`.

Stages 1 (time advance) and 3 (mass-property recompute) stay
inline in `step_internal` because each is < 5 lines and doesn't
benefit from a separate file. The detached-subtree ballistic
propagation also stays inline as it's a post-step concern, not
really part of stage 9.

## Verification

Mechanical refactor — no behavior change. All baselines bit-stable:

- `cargo build --workspace`, `fmt --check`, `clippy --workspace
  --tests -- -D warnings`, `RUSTDOCFLAGS=-D warnings cargo doc
  --workspace --no-deps`: clean
- `cargo nextest run --workspace -E 'not test(tier3_)'`:
  823/823 pass
- `cargo nextest run --workspace -E 'test(tier3_) and not
  test(earth_moon)'`: 308/308 pass
- `tier3_baseline_diff`: OK (78 matched; 0 missing; 0 new)

Closes the remaining stretch goal from #253.
