+++ crates/jeod_runner/src/simulation/step/integrate.rs
@@ -1,471 +1,32 @@
-//! Per-step integration pipeline for [`super::Simulation`].
-//!
-//! Houses `step`, `step_internal` (the ~1,000-line RK4 orchestrator
-//! covering JEOD's 9 integration stages), `step_n`, `step_until`,
-//! `frame_origin`, and `sync_pfix_rotation`. A future split per JEOD
-//! integration stage is tracked under issue #253 (Task B).
+//! Stages 7 + 8 + 8b of [`super::super::Simulation::step_internal`]:
+//! force collection, RK4 integration (with contact-pair coupling), and
+//! frame-switch handling. The bulk of `step_internal` lives here. The
+//! contact-pair branch interleaves stage-7 force collection with the
+//! per-RK4-stage derivative recomputation, so the three stages stay
+//! together rather than getting carved further.
 
 use glam::{DMat3, DVec3};
 
-use jeod_frames::FrameId;
-use jeod_sim::atmosphere::evaluate_atmosphere;
 use jeod_sim::forces::collect_and_resolve_forces;
 use jeod_sim::gravity::accumulate_gravity;
 use jeod_sim::integration::{integrate_bodies_contact_coupled, integrate_body, CoupledBodyInput};
 use jeod_sim::{
-    evaluate_contact_pair, integrate_body_coupled, CoupledStageEval, GravityControls, JeodQuat,
-    MassProperties, RadiationForce, RotationModel, RotationalState, SwitchSense,
-    TranslationalState,
+    evaluate_contact_pair, integrate_body_coupled, CoupledStageEval, GravityControls,
+    MassProperties, RadiationForce, RotationalState, SwitchSense, TranslationalState,
 };
 
-use super::types::ContactPairConfig;
-use super::Simulation;
+use super::super::types::ContactPairConfig;
+use super::super::Simulation;
 use crate::error::StepError;
 
 impl Simulation {
-    /// Advance the simulation by one timestep.
-    ///
-    /// Runs the full JEOD pipeline in order:
-    /// 1. Time update
-    /// 2. Ephemeris update (planet-fixed rotations + frame tree sync)
-    /// 3. Mass update (recompute derived quantities)
-    /// 4. Gravity computation
-    /// 5. Atmosphere evaluation
-    /// 6. Interaction computation (drag, SRP, gravity torque)
-    /// 7. Force collection and frame derivative computation
-    /// 8. State integration (RK4, with sub-stage tree updates)
-    /// 9. Derived state computation
-    pub fn step(&mut self) -> Result<(), StepError> {
-        self.step_internal(self.dt)
-    }
-
-    /// Get the position and velocity of a frame relative to the root inertial frame.
-    pub fn frame_origin(&self, frame_id: FrameId) -> (DVec3, DVec3) {
-        if frame_id == self.root_frame_id {
-            return (DVec3::ZERO, DVec3::ZERO);
-        }
-        let state = self
-            .frame_tree
-            .compute_relative_state(self.root_frame_id, frame_id);
-        (state.trans.position, state.trans.velocity)
-    }
-
-    /// Sync a planet-fixed frame node's rotation state from a computed matrix.
-    ///
-    /// Sets `t_parent_this`, derives `q_parent_this` from it, and sets
-    /// `ang_vel_this = [0, 0, planet_omega]` matching JEOD's `planet_rnp.cc`.
-    /// The `planet_omega` value comes from [`PlanetConfig::omega`] via
-    /// [`GravityData::planet_omega`].
-    fn sync_pfix_rotation(
-        frame_tree: &mut jeod_frames::FrameTree,
-        pfix_id: jeod_frames::FrameId,
-        rotation: DMat3,
-        planet_omega: f64,
-    ) {
-        let node = frame_tree.get_mut(pfix_id);
-        node.state.rot.t_parent_this = rotation;
-        node.state.rot.q_parent_this = JeodQuat::left_quat_from_transformation(&rotation);
-        // JEOD sets ang_vel_this = [0, 0, planet_omega] in planet_rnp.cc.
-        // This is used by compute_relative_state velocity composition.
-        node.state.rot.ang_vel_this = DVec3::new(0.0, 0.0, planet_omega);
-    }
-
-    /// Internal step with explicit dt (avoids temporary mutation of `self.dt`
-    /// in `step_until`).
-    fn step_internal(&mut self, dt: f64) -> Result<(), StepError> {
-        // ── 1. Time update ──
-        self.time.advance(dt);
-
-        // ── 2. Ephemeris update — planet-fixed rotations + frame tree sync ──
-        // JEOD_INV: DM.13 — ephemeris updated before gravity
-        // Per-source rotation dispatch: each source has its own rotation model.
-        // Lazy-compute Earth RNP only if needed (most common case).
-        let mut earth_rotation: Option<DMat3> = Option::None;
-        for (i, grav) in self.gravity_data.iter_mut().enumerate() {
-            match grav.rotation_model {
-                RotationModel::None => {}
-                RotationModel::EarthRNP => {
-                    let rotation = *earth_rotation.get_or_insert_with(|| {
-                        jeod_sim::compute_t_parent_this_from_tjt_with_polar(
-                            self.time.gmst_seconds,
-                            self.time.tt_tjt(),
-                            self.polar_motion,
-                        )
-                    });
-                    // Sync to frame tree pfix node.
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MarsIAU => {
-                    // JEOD's RNPMars receives TT seconds since J2000 (time_tt.seconds).
-                    let tt_s_since_j2000 = (self.time.tt_tjt() - jeod_time::epoch::J2000_TT_TJT)
-                        * jeod_time::epoch::SECONDS_PER_DAY;
-                    let rotation =
-                        jeod_frames::rotation_mars::compute_mars_rotation(tt_s_since_j2000);
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MoonIAU => {
-                    let tdb_jd = self.time.tdb_julian_date();
-                    let tdb_s_since_j2000 = (tdb_jd - jeod_time::epoch::J2000_TT_JD)
-                        * jeod_time::epoch::SECONDS_PER_DAY;
-                    let rotation =
-                        jeod_frames::rotation_moon::compute_moon_rotation(tdb_s_since_j2000);
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MoonDE421 => {
-                    let eph = self.ephemeris.as_ref().expect(
-                        "MoonDE421 rotation requires ephemeris with BPC. \
-                         Set sim.ephemeris = Some(eph) after calling eph.load_bpc().",
-                    );
-                    let tdb_jd = self.time.tdb_julian_date();
-                    let rotation = eph
-                        .get_body_rotation(jeod_sim::EphemerisBody::Moon, tdb_jd)
-                        .expect("Moon DE421 BPC rotation query failed");
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-            }
-            // Compute tidal ΔC20 if configured; otherwise clear any stale value.
-            if let Some(ref config) = grav.tidal_config {
-                let pfix_id = self.source_frame_ids[i]
-                    .pfix
-                    .expect("tidal_config requires a planet-fixed frame (set rotation_model or t_inertial_pfix).");
-                let rotation = self.frame_tree.get(pfix_id).state.rot.t_parent_this;
-                grav.delta_c20 = jeod_gravity::tides::compute_delta_c20(config, &rotation);
-            } else {
-                grav.delta_c20 = 0.0;
-            }
-        }
-
-        // ── 2b. Ephemeris update — source positions from DE4xx ──
-        // Update source positions from ephemeris each step and sync to frame tree.
-        if let Some(ref eph) = self.ephemeris {
-            let tdb_jd = self.time.tdb_julian_date();
-            for i in 0..self.source_ephem_bodies.len() {
-                if let Some(Some((target, observer))) = self.source_ephem_bodies.get(i) {
-                    let (pos_typed, vel_typed) = eph
-                        .get_state_typed(*target, *observer, tdb_jd)
-                        .map_err(|e| StepError::EphemerisLookup {
-                            source_idx: i,
-                            target: *target,
-                            observer: *observer,
-                            tdb_jd,
-                            message: e.to_string(),
-                        })?;
-                    let (pos, vel) = (pos_typed.raw_si(), vel_typed.raw_si());
-                    // Root-mapped sources cannot consume ephemeris position updates:
-                    // the root frame must remain identity, so accepting such a
-                    // mapping would silently ignore `pos` and yield an incorrect
-                    // source position.
-                    let fid = self.source_frame_ids[i].inertial;
-                    assert!(
-                        fid != self.root_frame_id,
-                        "Invalid ephemeris mapping for source {i} \
-                         ({target:?} wrt {observer:?}): source inertial frame is the root frame, \
-                         whose state must remain identity. Root-mapped sources cannot use \
-                         ephemeris position updates."
-                    );
-                    // Update frame tree node with ephemeris position/velocity.
-                    let node = self.frame_tree.get_mut(fid);
-                    node.state.trans.position = pos;
-                    node.state.trans.velocity = vel;
-                    // Also update gravity_data velocity for relativistic corrections.
-                    self.gravity_data[i].velocity = vel;
-                }
-            }
-        }
-
-        // ── 3. Mass update — recompute inverse_mass/inverse_inertia ──
-        for body in &mut self.bodies {
-            if let Some(ref mut mass) = body.mass {
-                mass.recompute_derived();
-            }
-        }
-
-        // Precompute frame origins from the tree for all body integration frames.
-        let body_integ_origins: Vec<(DVec3, DVec3)> = self
-            .bodies
-            .iter()
-            .map(|b| self.frame_origin(b.integ_frame_id))
-            .collect();
-
-        // ── 4. Environment — gravity ──
-        // Helper: resolve source to gravity data via frame tree.
-        let gravity_data = &self.gravity_data;
-        let source_frame_ids = &self.source_frame_ids;
-        let frame_tree = &self.frame_tree;
-        let root_fid = self.root_frame_id;
-        let resolve_source = |source_id: usize| -> Option<jeod_sim::ResolvedSource<'_>> {
-            let grav = gravity_data.get(source_id)?;
-            let sfids = &source_frame_ids[source_id];
-            let src_node = frame_tree.get(sfids.inertial);
-            let position = if sfids.inertial == root_fid {
-                DVec3::ZERO
-            } else {
-                src_node.state.trans.position
-            };
-            let rotation = sfids
-                .pfix
-                .map(|pfix_id| &frame_tree.get(pfix_id).state.rot.t_parent_this);
-            Some(jeod_sim::ResolvedSource {
-                source: &grav.source,
-                rotation,
-                position,
-                delta_c20: grav.delta_c20,
-                has_delta_coeffs: grav.tidal_config.is_some(),
-            })
-        };
-
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            let integ_origin = body_integ_origins[body_idx].0;
-            body.gravity_accel = accumulate_gravity(
-                body.trans.position + integ_origin,
-                &body.gravity_controls,
-                integ_origin,
-                resolve_source,
-            );
-        }
-
-        // ── 4b. Relativistic corrections ──
-        // After Newtonian gravity, apply post-Newtonian PPN correction for
-        // any source with `relativistic: true`. Folkner eq 27 (β=γ=1).
-        // PPN uses inertial coordinates — convert from integration frame.
-        let resolve_rel_source =
-            |source_id: usize| -> Option<jeod_sim::ResolvedRelativisticSource> {
-                let grav = gravity_data.get(source_id)?;
-                let sfids = &source_frame_ids[source_id];
-                let position = if sfids.inertial == root_fid {
-                    DVec3::ZERO
-                } else {
-                    frame_tree.get(sfids.inertial).state.trans.position
-                };
-                Some(jeod_sim::ResolvedRelativisticSource {
-                    mu: grav.source.mu,
-                    position,
-                    // Use velocity from gravity_data, not the tree node, because
-                    // central bodies at the root frame have zero tree velocity
-                    // but may have physical velocity for relativistic corrections.
-                    velocity: grav.velocity,
-                })
-            };
-
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            let (origin, origin_vel) = body_integ_origins[body_idx];
-            body.gravity_accel.grav_accel += jeod_sim::accumulate_relativistic_corrections(
-                body.trans.position + origin,
-                body.trans.velocity + origin_vel,
-                &body.gravity_controls,
-                resolve_rel_source,
-            );
-        }
-
-        // ── 5. Environment — atmosphere ──
-        if let Some(ref atmos_config) = self.atmosphere {
-            let t_pfix = self
-                .atmosphere_planet_source
-                .and_then(|idx| self.source_frame_ids.get(idx))
-                .and_then(|sfids| sfids.pfix)
-                .map(|pfix_id| &self.frame_tree.get(pfix_id).state.rot.t_parent_this);
-            let tai_tjt = Some(self.time.tai_tjt);
-
-            for body in &mut self.bodies {
-                if body.atmospheric_state.is_some() {
-                    body.atmospheric_state = Some(evaluate_atmosphere(
-                        atmos_config,
-                        body.trans.position,
-                        t_pfix,
-                        tai_tjt,
-                    ));
-                }
-            }
-        }
-
-        // ── 6. Interactions — drag, SRP, gravity torque ──
-        // sun_pos is also used in stage 9 (solar beta, earth lighting); compute once here.
-        let sun_pos = self.sun_source.map(|idx| self.source_position(idx));
-        let moon_pos = self.moon_source.map(|idx| self.source_position(idx));
-        let source_frame_ids = &self.source_frame_ids;
-        let frame_tree = &self.frame_tree;
-        let root_fid = self.root_frame_id;
-
-        for body in &mut self.bodies {
-            // Compute structural transform once (shared by drag and flat-plate SRP)
-            let t_inertial_body = body.rot.as_ref().map_or(DMat3::IDENTITY, |r| {
-                r.quaternion.left_quat_to_transformation()
-            });
-            let t_inertial_struct =
-                jeod_sim::compute_t_inertial_struct(&body.t_struct_body, &t_inertial_body);
-
-            // Aerodynamic drag
-            body.aero_force = None;
-            if let (Some(ref drag_config), Some(ref atmos)) = (&body.drag, &body.atmospheric_state)
-            {
-                body.aero_force = Some(jeod_sim::compute_drag(
-                    drag_config,
-                    atmos,
-                    body.trans.velocity,
-                    body.rot.as_ref(),
-                    body.t_struct_body,
-                ));
-            }
-
-            // Solar radiation pressure (flat-plate)
-            body.radiation_force = None;
-            if let Some(ref mut fps) = body.flat_plate_state {
-                fps.stage_inputs = None;
-            }
-            if let Some(sun_position) = sun_pos {
-                if let Some(ref mut fps) = body.flat_plate_state {
-                    // Flat-plate SRP with thermal emission
-                    let sun_to_vehicle = body.trans.position - sun_position;
-                    let distance = sun_to_vehicle.length();
-                    // Skip SRP (not the whole body) if too close to Sun
-                    if distance >= 1.0 {
-                        let flux_inertial_hat = sun_to_vehicle / distance;
-                        let flux_mag = jeod_sim::solar_flux_at_distance(distance);
-
-                        // Shadow fraction (step-constant — matches JEOD's
-                        // scheduled-class shadow evaluation in SIM_3_ORBIT).
-                        let illum_factor = body
-                            .shadow_body
-                            .map(|(idx, radius)| {
-                                jeod_sim::compute_shadow_fraction(
-                                    body.trans.position,
-                                    sun_position,
-                                    {
-                                        let fid = source_frame_ids[idx].inertial;
-                                        if fid == root_fid {
-                                            DVec3::ZERO
-                                        } else {
-                                            frame_tree.get(fid).state.trans.position
-                                        }
-                                    },
-                                    radius,
-                                    jeod_sim::SOLAR_RADIUS,
-                                )
-                            })
-                            .unwrap_or(1.0);
-
-                        let center_grav = body.mass.as_ref().map_or(DVec3::ZERO, |m| m.position);
-
-                        match fps.integration_order {
-                            jeod_sim::ThermalIntegrationOrder::Scheduled => {
-                                // Scheduled-class: compute SRP force + Euler T
-                                // update once per step (JEOD SIM_3_ORBIT).
-                                let flux_struct_hat = t_inertial_struct * flux_inertial_hat;
-                                let srp_result = jeod_sim::compute_flat_plate_srp_thermal(
-                                    &fps.plates,
-                                    &fps.t_pow4_cached,
-                                    flux_struct_hat,
-                                    flux_mag,
-                                    center_grav,
-                                    illum_factor,
-                                );
-
-                                // Force: structural → inertial. Torque: stays structural.
-                                let force_inertial =
-                                    t_inertial_struct.transpose() * srp_result.force;
-                                body.radiation_force = Some(RadiationForce {
-                                    force: force_inertial,
-                                    torque: srp_result.torque,
-                                });
-
-                                fps.integrate_temperatures(&srp_result.temp_dots, dt);
-                            }
-                            jeod_sim::ThermalIntegrationOrder::DerivativeFirstOrder
-                            | jeod_sim::ThermalIntegrationOrder::DerivativeRk4 => {
-                                // Derivative-class: SRP force (and optionally T)
-                                // recomputed per RK4 stage. Stash the step-start
-                                // inputs on the plate state; Stage 8 consumes
-                                // them via `integrate_body_coupled` below.
-                                fps.stage_inputs = Some(jeod_sim::FlatPlateStageInputs {
-                                    sun_position,
-                                    illum_factor,
-                                    center_grav,
-                                });
-                                // `radiation_force` is left None here; Stage 8
-                                // writes a representative final-stage value so
-                                // `VehicleOutput` still reports SRP force.
-                            }
-                        }
-                    }
-                } else if let Some((cx_area, albedo, diffuse)) = body.cannonball_srp {
-                    let illum_factor = body
-                        .shadow_body
-                        .map(|(idx, radius)| {
-                            jeod_sim::compute_shadow_fraction(
-                                body.trans.position,
-                                sun_position,
-                                {
-                                    let fid = source_frame_ids[idx].inertial;
-                                    if fid == root_fid {
-                                        DVec3::ZERO
-                                    } else {
-                                        frame_tree.get(fid).state.trans.position
-                                    }
-                                },
-                                radius,
-                                jeod_sim::SOLAR_RADIUS,
-                            )
-                        })
-                        .unwrap_or(1.0);
-
-                    let force = jeod_sim::compute_cannonball_srp(
-                        body.trans.position,
-                        sun_position,
-                        cx_area,
-                        albedo,
-                        diffuse,
-                        illum_factor,
-                    );
-                    if force != DVec3::ZERO {
-                        body.radiation_force = Some(RadiationForce {
-                            force,
-                            torque: DVec3::ZERO,
-                        });
-                    }
-                }
-            }
-
-            // Gravity gradient torque
-            body.gravity_torque = None;
-            if body.compute_gravity_torque {
-                if let (Some(ref rot), Some(ref mass)) = (&body.rot, &body.mass) {
-                    body.gravity_torque = Some(jeod_sim::compute_gravity_torque(
-                        &body.gravity_accel.grav_grad,
-                        rot,
-                        &mass.inertia,
-                    ));
-                }
-            }
-        }
-
+    /// Stages 7 + 8 + 8b — force collection, RK4 integration, and
+    /// frame-switch handling.
+    pub(super) fn run_integration(
+        &mut self,
+        dt: f64,
+        body_integ_origins: &[(DVec3, DVec3)],
+    ) -> Result<(), StepError> {
         // ── 7. Force collection ──
         for body in &mut self.bodies {
             let (total, derivs) = collect_and_resolve_forces(
@@ -1051,142 +612,6 @@ impl Simulation {
             }
         }
 
-        // ── 9. Derived states ──
-        let gravity_data = &self.gravity_data;
-
-        for body in &mut self.bodies {
-            // Orbital elements
-            if let Some(src_idx) = body.orbital_elements_source {
-                if let Some(mu) = gravity_data.get(src_idx).map(|g| g.source.mu) {
-                    body.orbital_elements = jeod_sim::compute_orbital_elements(
-                        mu,
-                        body.trans.position,
-                        body.trans.velocity,
-                    )
-                    .ok();
-                } else {
-                    body.orbital_elements = None;
-                }
-            }
-
-            // Euler angles
-            if let Some(seq) = body.euler_sequence {
-                if let Some(ref rot) = body.rot {
-                    body.euler_angles = Some(jeod_sim::compute_body_euler_angles(rot, seq));
-                } else {
-                    body.euler_angles = None;
-                }
-            }
-
-            // LVLH frame
-            if body.compute_lvlh {
-                body.lvlh_frame = Some(jeod_sim::compute_body_lvlh_frame(
-                    body.trans.position,
-                    body.trans.velocity,
-                ));
-            }
-
-            // Geodetic state
-            if let Some((src_idx, r_eq, r_pol)) = body.geodetic_planet {
-                let pfix_rot = self
-                    .source_frame_ids
-                    .get(src_idx)
-                    .and_then(|sfids| sfids.pfix)
-                    .map(|pfix_id| self.frame_tree.get(pfix_id).state.rot.t_parent_this);
-                if let Some(t_pfix) = pfix_rot {
-                    body.geodetic_state = Some(jeod_sim::compute_body_geodetic(
-                        body.trans.position,
-                        &t_pfix,
-                        r_eq,
-                        r_pol,
-                    ));
-                } else {
-                    body.geodetic_state = None;
-                }
-            }
-
-            // Solar beta
-            if body.compute_solar_beta {
-                if let Some(sp) = sun_pos {
-                    body.solar_beta = Some(jeod_sim::compute_body_solar_beta(
-                        body.trans.position,
-                        body.trans.velocity,
-                        sp,
-                    ));
-                } else {
-                    body.solar_beta = None;
-                }
-            }
-
-            // Earth lighting
-            if let Some((earth_r, moon_r, sun_r)) = body.earth_lighting_config {
-                if let (Some(sp), Some(mp)) = (sun_pos, moon_pos) {
-                    body.earth_lighting =
-                        Some(jeod_interactions::earth_lighting::compute_earth_lighting(
-                            body.trans.position,
-                            sp,
-                            mp,
-                            sun_r,
-                            earth_r,
-                            moon_r,
-                        ));
-                } else {
-                    body.earth_lighting = None;
-                }
-            }
-        }
-
-        // Advance any free-flying detached subtrees ballistically. This
-        // matches JEOD's behavior for tree roots whose grav_interaction
-        // is empty (the common case for staging — no force applied to
-        // separated stages between detach and reattach). Use the dynamic
-        // timestep `dt * time_scale_factor` so the ballistic propagation
-        // stays consistent with the integrated bodies under time reversal
-        // / scaling (matches `integ_dt` used elsewhere in this step).
-        if !self.detached_subtrees.is_empty() {
-            self.step_detached_subtrees(dt * self.time.time_scale_factor);
-        }
-
-        Ok(())
-    }
-
-    /// Advance the simulation by `n` timesteps.
-    pub fn step_n(&mut self, n: usize) -> Result<(), StepError> {
-        for _ in 0..n {
-            self.step()?;
-        }
-        Ok(())
-    }
-
-    /// Advance the simulation until `target_time` (in simulation seconds).
-    ///
-    /// Steps at `self.dt` until the remaining time is less than `dt`,
-    /// then takes a final fractional step if the remainder exceeds 1 ms.
-    pub fn step_until(&mut self, target_time: f64) -> Result<(), StepError> {
-        while self.time.simtime + self.dt <= target_time + 0.001 {
-            self.step()?;
-        }
-        let remainder = target_time - self.time.simtime;
-        if remainder > 0.001 {
-            // Fractional steps corrupt multi-step history arrays (GJ's
-            // Störmer-Cowell coefficients and delinv accumulators, ABM4's
-            // Adams history) — both methods assume constant dt.
-            let has_multistep = self.bodies.iter().any(|b| {
-                matches!(
-                    b.integrator,
-                    jeod_dynamics::IntegratorType::GaussJackson(..)
-                        | jeod_dynamics::IntegratorType::Abm4
-                )
-            });
-            assert!(
-                !has_multistep,
-                "step_until() would take a fractional step ({remainder:.6}s vs dt={:.6}s). \
-                 Multi-step integrators (GaussJackson, ABM4) require constant dt. \
-                 Ensure target_time is an integer multiple of dt.",
-                self.dt
-            );
-            self.step_internal(remainder)?;
-        }
         Ok(())
     }
 }
